src/config/requirements/development.txt
-r base.txt
django-debug-toolbar>=1.8
Werkzeug>=0.12.2
psycopg2